minor improvements from Andrew

+++ cinelerra-5.1/guicast/vframe.C
@@ -1403,12 +1403,12 @@ int VFrame::get_memory_usage()
 // a (~alpha) transparency, 0x00==solid .. 0xff==transparent
 void VFrame::set_pixel_color(int rgb, int a)
 {
-	pixel_rgb = (~a<<24) | (rgb&0xffffff);
+	pixel_rgb = (~(unsigned int)a<<24) | (rgb&0xffffff);
 	int ir = 0xff & (pixel_rgb >> 16);
 	int ig = 0xff & (pixel_rgb >> 8);
 	int ib = 0xff & (pixel_rgb >> 0);
 	YUV::yuv.rgb_to_yuv_8(ir, ig, ib);
-	pixel_yuv = (~a<<24) | (ir<<16) | (ig<<8) | (ib<<0);
+	pixel_yuv = (~(unsigned int)a<<24) | (ir<<16) | (ig<<8) | (ib<<0);
 }
 
 void VFrame::set_stiple(int mask)
